% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{coverage-class}
\alias{coverage-class}
\title{Create an S4 class for coverages}
\description{
Create an S4 class for coverages

Create an S4 class for coverages
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the dataframe that the coverage was generated from}

\item{\code{spatial}}{a list containing all the spatial coverage data that was extracted.  This will be a bounding box or a gpolygon}

\item{\code{temporal}}{a list of extracted temporal coverage data, usually just a begin and end date, but could be multiple start and end dates.}

\item{\code{taxanomic}}{a list containing the taxanomic coverage metadata}

\item{\code{name}}{the name of the dataframe that the coverage was generated from}

\item{\code{spatial}}{a list containing all the spatial coverage data that was extracted.  This will be a bounding box or a gpolygon}

\item{\code{temporal}}{a list of extracted temporal coverage data, usually just a begin and end date, but could be multiple start and end dates.}

\item{\code{taxanomic}}{a list containing the taxanomic coverage metadata}
}}

