% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coverage}
\alias{coverage}
\title{Coverage: The coverage data model}
\usage{
coverage()
}
\description{
This function creates a coverage data object.
}
\details{
The coverage object stores metadata created
by other functions in this package
}
\examples{
new_coverage <- coverage()
}

