% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{coverage_extract-method}
\alias{coverage_extract}
\alias{coverage_extract,data.frame-method}
\alias{coverage_extract-method}
\title{Extract coverages from a dataset.}
\usage{
coverage_extract(df, spatial = NULL, temporal = NULL, taxanomxic = NULL)

\S4method{coverage_extract}{data.frame}(df, spatial, temporal)
}
\arguments{
\item{df}{The dataset to extract coverages from, this should be a dataframe.}

\item{spatial}{a list of control parameters for extracting spatial coverages, see details for a full description}
}
\value{
Returns a coverage object with all coverages for a given dataset.
}
\description{
Extract temporal, spatial and taxanomic coverages from a dataset
}
\details{
Eeach coverage needs a corresponding list of parameters that specificy where the extractor can look for the appropriate data.
\itemize{
\item {spatial} {a list that contains three control parameters}}
}
\examples{
\dontrun{

Examp goes here
}
}

