% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datetime_range}
\alias{datetime_range}
\title{Automatically calculate start/end datetime range for your data}
\usage{
datetime_range(datetimes)
}
\arguments{
\item{datetimes}{A vector of \code{POSIXt} or \code{date}}
}
\value{
A vector of length 2 representing the first and last datetime
}
\description{
Take a vector of datetimes and calculate the range
}

