% Generated by roxygen2 (4.0.1): do not edit by hand
\name{points_to_poly}
\alias{points_to_poly}
\title{create polygon from series of points}
\usage{
points_to_poly(points, shape = "convex")
}
\arguments{
\item{points}{a matrix of latitute and longitude points, or a data.frame with
\code{latitude} and \code{longitude} as column names}

\item{shape}{a string specifying the shape of the polygon, defaults as
\code{'convex'}. Can be either \code{'convex'} or \code{'concave'}}
}
\value{
a data.frame of latitude and longitude points representing the spatial footprint of the data
}
\description{
create polygon from series of points
}
\details{
creates a automatically generated footprint from a series of points
}
\examples{
example_points <- read.table(file='../example_data.tsv',header=T,sep='\\t')
point_matrix <- matrix(as.numeric(c(example_points[, 2],example_points[, 1])),nrow=nrow(example_points),ncol=2)
points_to_poly(point_matrix)
}
\author{
Jordan S. Read
}
\keyword{math}
\keyword{methods,}

