% Generated by roxygen2 (4.0.1): do not edit by hand
\name{taxonomy_handler}
\alias{taxonomy_handler}
\title{This function creates a taxonomic coverage data object.}
\usage{
taxonomy_handler(input, genus_epithet = NULL, genus = NULL,
  epithet = NULL, subspecies = NULL)
}
\description{
This function creates a taxonomic coverage data object.
}
\examples{
\dontrun{
datpath <- system.file("inst/ext/data_gbif_1.csv", package = "mdextract")
dat <- read.csv(datpath)
taxonomy_handler(input=dat, fields=c('name'))
}
}

