% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bounding_box}
\alias{bounding_box}
\title{automatically generate coverages for a spocc search}
\usage{
bounding_box(lonlat)
}
\arguments{
\item{lonlat}{a 2 x n matrix of longitude-latitude coordinates}
}
\description{
This function will automatically generate metadata for spocc queries that can then be converted to other standards.
}

