% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datetime_range}
\alias{datetime_range}
\title{Automatically calculate start/end datetime range for your data}
\usage{
datetime_range(datetimes, type = "simple")
}
\arguments{
\item{datetimes}{A vector of \code{POSIXt} or \code{date}}
}
\value{
A data.frame with columns \code{starts} and \code{ends}.
Type \code{simple} always returns only one start/end pair.
}
\description{
Take a vector of datetimes and calculate the range
}

