% Generated by roxygen2 (4.0.1): do not edit by hand
\name{points_to_poly}
\alias{points_to_poly}
\title{create polygon from series of points}
\usage{
points_to_poly(pts, method = "convex")
}
\arguments{
\item{pts}{a matrix of latitute and longitude points, or a data.frame with
\code{latitude} and \code{longitude} as column names}

\item{method}{a string specifying the shape of the polygon, defaults as
\code{'convex'}. Currently \code{'convex'} is the only supported method.}
}
\value{
a data.frame of latitude and longitude points representing the spatial footprint of the data
}
\description{
create polygon from series of points
}
\details{
creates a automatically generated footprint from a series of points
}
\examples{
file_nm <- system.file(package='mdextract','ext','data_gbif_1.csv')
example_points <- read.table(file=file_nm,header=T,sep=',')
point_matrix <- matrix(as.numeric(c(example_points$decimalLongitude,example_points$decimalLatitude)),
 nrow=nrow(example_points),ncol=2)
poly <- points_to_poly(point_matrix)
plot(point_matrix)
plot_poly(poly)
}
\author{
Jordan S. Read
}
\keyword{math}
\keyword{methods,}

